requirements.txt
torch
llama_cpp_python
huggingface-hub
sentencepiece
transformers